Refactor GymDetailVC to use FacilityDropdown
